moved get service extras to registry proxy
