+++ services/director/src/simcore_service_director/producer.py
@@ -18,10 +18,6 @@ from . import config, docker_utils, exceptions, registry_proxy
 from .config import APP_CLIENT_SESSION_KEY
 from .system_utils import get_system_extra_hosts_raw
 
-SERVICE_RUNTIME_SETTINGS: str = "simcore.service.settings"
-SERVICE_REVERSE_PROXY_SETTINGS: str = "simcore.service.reverse-proxy-settings"
-SERVICE_RUNTIME_BOOTSETTINGS: str = "simcore.service.bootsettings"
-
 log = logging.getLogger(__name__)
 
 
@@ -84,7 +80,7 @@ def _parse_mount_settings(settings: List[Dict]) -> List[Dict]:
         log.debug("Append mount settings %s", mount)
         mounts.append(mount)
 
-    return mount
+    return mounts
 
 
 def _parse_env_settings(settings: List[str]) -> Dict:
@@ -128,10 +124,10 @@ async def _create_docker_service_params(
 ) -> Dict:
     # pylint: disable=too-many-statements
     service_parameters_labels = await _read_service_settings(
-        app, service_key, service_tag, SERVICE_RUNTIME_SETTINGS
+        app, service_key, service_tag, config.SERVICE_RUNTIME_SETTINGS
     )
     reverse_proxy_settings = await _read_service_settings(
-        app, service_key, service_tag, SERVICE_REVERSE_PROXY_SETTINGS
+        app, service_key, service_tag, config.SERVICE_REVERSE_PROXY_SETTINGS
     )
     service_name = registry_proxy.get_service_last_names(service_key) + "_" + node_uuid
     log.debug("Converting labels to docker runtime parameters")
@@ -666,7 +662,7 @@ async def _start_docker_service(
         published_port, target_port = await _get_docker_image_port_mapping(service)
         # now pass boot parameters
         service_boot_parameters_labels = await _read_service_settings(
-            app, service_key, service_tag, SERVICE_RUNTIME_BOOTSETTINGS
+            app, service_key, service_tag, config.SERVICE_RUNTIME_BOOTSETTINGS
         )
         service_entrypoint = _get_service_entrypoint(service_boot_parameters_labels)
         if published_port:
@@ -830,7 +826,7 @@ async def _get_node_details(
     # get boot parameters
     results = await asyncio.gather(
         _read_service_settings(
-            app, service_key, service_tag, SERVICE_RUNTIME_BOOTSETTINGS
+            app, service_key, service_tag, config.SERVICE_RUNTIME_BOOTSETTINGS
         ),
         _get_service_basepath_from_docker_service(service),
         _get_service_state(client, service),
@@ -996,52 +992,3 @@ async def stop_service(app: web.Application, node_uuid: str) -> None:
                 "DYNAMIC",
                 "SUCCESS",
             )
-
-
-ORG_LABELS_TO_SCHEMA_LABELS = {
-    "org.label-schema.build-date": "build_date",
-    "org.label-schema.vcs-ref": "vcs_ref",
-    "org.label-schema.vcs-url": "vcs_url",
-}
-
-
-async def generate_service_extras(
-    app: web.Application, image_key: str, image_tag: str
-) -> Dict:
-    result = {}
-    labels = await registry_proxy.get_image_labels(app, image_key, image_tag)
-    log.debug("Compiling service extras from labels %s", labels)
-
-    # check physical node requirements
-    # all nodes require "CPU"
-    result["node_requirements"] = ["CPU"]
-    # check if the service requires GPU support
-
-    def validate_kind(entry_to_validate, kind_name):
-        for element in (
-            entry_to_validate.get("value", {})
-            .get("Reservations", {})
-            .get("GenericResources", [])
-        ):
-            if element.get("DiscreteResourceSpec", {}).get("Kind") == kind_name:
-                return True
-        return False
-
-    if SERVICE_RUNTIME_SETTINGS in labels:
-        service_settings = json.loads(labels[SERVICE_RUNTIME_SETTINGS])
-        for entry in service_settings:
-            if entry.get("name") == "Resources" and validate_kind(entry, "VRAM"):
-                result["node_requirements"].append("GPU")
-            if entry.get("name") == "Resources" and validate_kind(entry, "MPI"):
-                result["node_requirements"].append("MPI")
-
-    # get org labels
-    result.update(
-        {
-            sl: labels[dl]
-            for dl, sl in ORG_LABELS_TO_SCHEMA_LABELS.items()
-            if dl in labels
-        }
-    )
-
-    return result
