+++ services/director/tests/test_registry_proxy.py
@@ -266,3 +266,24 @@ async def test_get_services_performance(
             (stop_time - start_time) / len(services),
         )
     )
+
+
+async def test_generate_service_extras(
+    aiohttp_mock_app,
+    push_services,
+    configure_registry_access,
+    configure_schemas_location,
+):
+    images = push_services(
+        number_of_computational_services=1, number_of_interactive_services=1
+    )
+
+    for image in images:
+        service_description = image["service_description"]
+        service_extras = image["service_extras"]
+
+        extras = await registry_proxy.get_service_extras(
+            aiohttp_mock_app, service_description["key"], service_description["version"]
+        )
+
+        assert extras == service_extras
